requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
murf==1.2.2
pydub==0.25.1
requests==2.31.0
google-generativeai==0.3.2
python-pptx==0.6.23
PyPDF2==3.0.1
youtube-transcript-api==1.0.3
SpeechRecognition==3.10.1
defusedxml==0.7.1
gunicorn==21.2.0
google-cloud-storage==2.16.0
moviepy==1.0.3